requirements.txt
graphql-core
strawberry-graphql
aiohttp
gql
requests